ninja-api/dveapi/pyproject.toml
[tool.poetry]
name = "dveapi"
version = "0.1.0"
description = ""
authors = ["Caspar Trittibach <caspar.trittibach@haag-streit.com>"]
readme = "README.md"
packages = [
    { include = "cli"},
    { include = "config"},
    { include = "authenticate"},
    { include = "fileManager"},
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
uvicorn = "^0.24.0.post1"
django-ninja = "^1.0.1"
django-ninja-extra = "^0.20.0"
ninja-schema = "^0.13.6"
django-ninja-jwt = "^5.3.0"
django-cors-headers = "^4.3.1"
toml = "^0.10.2"
requests = "^2.31.0"
logging = "^0.4.9.6"
psycopg2-binary = "^2.9.9"

[tool.poetry.scripts]
djm = 'cli.manage:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
log_cli = 1
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"


